Do not check active endpoints
Registering an already active endpoint returns a non-zero status code in
the AfRegisterResponse, but is otherwise harmless. However, a new
version of the firmware does not support querying the active endpoints,
causing a delay waiting for the ZdoActiveEP callback to time out.
Therefore, it is best to not check the active endpoints and always try
to register all endpoints.